+++ R/db_bioinfo.R
@@ -145,15 +145,16 @@ getMetadataGeneSetTable <- function() {
 
 #' @rdname getGeneSet
 #' @export
+getAvailableGeneSets <- function(){
+  sql <- "SELECT DISTINCT genesetname FROM public.geneassignment"
+  getPostgresql(sql)[["genesetname"]]
+}
+
+#' @rdname getGeneSet
+#' @export
 getGeneSet <- function(geneset) {
-  
   sql <- glue::glue_sql(
-    "SELECT 
-       ensg
-     FROM
-       public.geneassignment
-     WHERE 
-       genesetname = {geneset}",
+    "SELECT ensg FROM public.geneassignment WHERE genesetname = {geneset}",
     .con = DBI::ANSI()
   )